src/python/api/inference/requirements.txt
black
flake8
hurry.filesize==0.9
ultralytics
yt-dlp
youtube-dl==2020.12.2
opencv-python
pafy
cap_from_youtube
fastapi
azure-storage-blob
azure-identity
python-multipart
pydantic